renamed transcribe.py to Tesseract.py and updated processing/transcription function for 1 image
